function 2 done right

+++ source/functions.c
@@ -48,7 +48,7 @@ bool EndOfFile(char character) {
 float CalculateTFIDF(int frequency, int documentQuantity, int wordAppearance) {
     float tf = frequency;
 
-    float x = (1 + documentQuantity) / (1 + wordAppearance);
+    float x = (1 + (float)documentQuantity) / (1 + (float)wordAppearance);
     float idf = log(x) + 1;
 
     return tf * idf;
